(chore) move NewFieldTemplate folder into logbooks/
