+++ client/src/components/logbooks/fields/NewItemTemplate/NewItemTemplate.tsx
@@ -1,7 +1,7 @@
-import NewFieldTemplate from "@/components/logbooks/fields/NewFieldTemplate/NewFieldTemplate";
 import MiniField from "@/components/logbooks/fields/NewItemTemplate/MiniField";
 import useNewItemTemplate from "@/components/logbooks/fields/NewItemTemplate/useNewItemTemplate";
 import F from "@/components/logbooks/LogbookForm/style/LogbookForm.style";
+import NewFieldTemplate from "@/components/logbooks/NewFieldTemplate/NewFieldTemplate";
 import { font } from "@/lib/theme/font";
 import type { ID } from "@t/data/utility.types";
 import { LucideBookCopy } from "lucide-react";
